+++ d4np/d4np_c.h
@@ -29,6 +29,7 @@
 /* Generic data structures */
 #include "d4np/ds/hashmap.h"
 #include "d4np/ds/linked_list.h"
+#include "d4np/ds/ring_buffer.h"
 #include "d4np/ds/string_builder.h"
 #include "d4np/ds/vector.h"